Issue 1257: Minor refactoring of the strongbox-data-service module
