+++ strongbox-data-service/src/main/java/org/carlspring/strongbox/config/orientdb/CommonOrientDbConfig.java
@@ -1,4 +1,4 @@
-package org.carlspring.strongbox.config;
+package org.carlspring.strongbox.config.orientdb;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
